refactor writing of notes
