+++ tests/test_features_notes.py
@@ -12,7 +12,8 @@ from os.path import join as pj
 
 from madmom.features import Activations
 from madmom.features.notes import *
-from madmom.io import expand_notes, load_notes, write_notes, write_mirex_format
+from madmom.io import load_notes, write_notes, write_notes_mirex
+from madmom.utils import expand_notes
 from . import ACTIVATIONS_PATH, ANNOTATIONS_PATH, AUDIO_PATH
 
 sample_file = pj(AUDIO_PATH, "stereo_sample.wav")
@@ -60,7 +61,7 @@ class TestWriteNotesFunction(unittest.TestCase):
 class TestWriteMirexFormatFunction(unittest.TestCase):
 
     def test_values(self):
-        result = write_mirex_format(
+        result = write_notes_mirex(
             NOTES, pj(ANNOTATIONS_PATH, 'stereo_sample.notes.mirex'))
         self.assertTrue(np.allclose(result[:, 0], NOTES[:, 0]))
         self.assertTrue(np.allclose(result[:, 1], NOTES[:, 0] + NOTES[:, 2]))
